Assignment_3/Q3/requirements.txt
numpy==1.24.4
matplotlib==3.7.5
PyYAML==6.0.2
tqdm==4.67.1
plyfile==1.0.3
wandb==0.17.9
pandas==2.0.3

# PyTorch (Py3.8-compatible builds)
torch==2.4.1
torchvision==0.19.1